EN/EF: Taking Chuck's Topaz changeset 6f4be60003aa: Updated the multi-compile script to assert if the objects don't match the source file names or if the object directory was not $(o).
From https://www.yammer.com/bentley.com/#/Threads/show?threadId=426303402:
Multi-compile makefile change:
I have pushed a change to MultiCppCompileRule.mki that may stop your builds. During multi-compile we use the list of objects and sources you provide to determine if we need to compile the sources; however when we do the actual compile the object files must match the source files. There is no list of OBJs passed to CL. Furthermore we make the assumption that all the objects must end up in $(o); that's the single object directory passed to CL. If either of these is not true in your makefiles then they will rebuild unneccessarily. Fixing these will not change the way that things are built; it will just improve the performance on an incremental compile. I have used the new ~assert capability in bmake to alert folks when either of these conditions is not met. It is in the Topaz stream and it should go into Vancouver at the beginning of next week. I have already fixed up everything that is built with MicroStation, and it did not take long using "bb b -p" to continue once the error was fixed.
What do you need to do?
If you get an error like "Multi-compiled files must compile to objects with the same name. Failing on dependency ECXAttrSessionManager.cpp", you should go to the makefile, look at the line with that dependency, and make sure the object file is the same as the source. Sometimes it's just a case error, other times it's completely wrong because of a copy/paste error of the previous line. Either way, copy the source file name on top of the object file name and continue. (If you really need an object to have a different name than the source, then it cannot participate in multi-compile. But remember that it is currently going to the same name.)
If you get an error like "Multi-compiled objects must go to $(o) as their object directory. Failing on dependency ECXAttrSessionManager.cpp" it's a little more complicated. Most likely you have defined your own object directory (i.e. "$(objDir)foo.obj : $(baseDir)foo.cpp"). You can either replace your object directory in those lines with $(o) as in "$(o)foo.obj : $(baseDir)foo.cpp", or set "o=$(objDir)" before the multi-compile. The preferred solution depends on why you supplied your own object dir. Remember that everything was already going to $(o) regardless of what you specified.