Changed VisEdgesLib to export only one mechanism for generating hidden lines: VisibleEdgesProcessor. This is a subclass of ViewContext. The caller is responsible for calling the _VisitElement method on this context for each element of interest.
VisEdgesLib now delay-loads PSolidCore. This is to allow applications such as SampleWallsEditor to link with and call functions in
VisEdgesLib, even if PSolidCOre and PSolid are not available. VisibleEdgesProcessor fails gracefully if Parasolid is not available.

Added a new method to ViewController: _VisitElements. This calls the _VisitElement on the supplied ViewContext for *every*
element in the view, not just the subset of elements that are displayed.

SampleWallsEditor now optionally sets up the environment to add the location of PSolidAddin to the path, so that Parasolid can be loaded.