Removed ECRelationshipConstraintClass and key properties. ECRelationshipConstraint code and documentation clean-up.
The ECRelationshipConstraintClass served the purposed of holding key properties. Since they have now been removed there is no need for the extra layer of abstraction on top of ECClass for
constraint classes. ECRelationshipClassList now holds a bvector of ECClassCPs instead of the using a std:vector of ECRelationshipConstraintClasses.

The documentation of both ECRelationshipClassList and ECRelationshipConstraint has been updated to more accurately describe what is happening.