TFS#630461: Make a PrimitiveArrayECProperty unable to override a PrimitiveECProperty
Previously, a PrimivitiveArrayECProperty could override a PrimitiveECProperty, or vice-versa, for a handful of schemas. With changes to the conversion of EC2.0/3.0
schemas, any schema which was violating the override was being automatically renamed.

The ability to override has now been removed from all schemas and is fixed in the handful of schemas which are 'allowed' to violate the rule. It has been fixed by changing
all PrimitiveECProperties, that violate the rule, to PrimitiveArrayECProperties during deserialization, this includes all derived properties of the converted property.