EF: __PUBLISH_CLASS_VIRTUAL__ Is Now An Error On bim0200dev
Use of __PUBLISH_CLASS_VIRTUAL__ has been deprecated on bim0200dev for some time now, and occurrences used to be flagged as warnings. I think I have cleaned the remnants in DgnClientSdk and turned usage of this indicator into an error.

Generally speaking you should simply remove this line in any header file that still has it. Note that the only useful thing PublishApi.py still did with this marker was to leave "public:" at the end of the replacement, so if this line was not immediately followed by public: anyway (typical pattern), you should add it.