review tests for YawPitchRoll drift. These tests have previously shown two "problems" 1) several pages of warnings in console output 2) intermittent hard failure in PRG builds
for Problem 1: The output occurs when an extremely tight tolerance condition fails.
The tight condition is 100X smaller than the hard failure tolerance.
So the test code is altered to only do warnings for 10X smaller.

for Problem 2: The PRG logs for the failed case give YPR angles for the failure.
Feeding these as a separate test confirms that they produce good results.

The whole test appears to work fine.