BF: remove ReadFileProperties() fallback to instance label for file name
This caused issues for classes that were not meant to have file name in their
properties and label included user defined string with characters invalid for file
path. This should just fallback to generated name.

If we need support for labeled class - either schema needs to be changed (supplementing it in client)
or approach redesigned.