Add test to evaluate atan2(-0.0, -1), atan2(+0.0, 1.0), Angle::Atan (-0.0,-1.0), Angle (0.0, -1.0).
Angle::Atan2 modified to test if the first arg is -0.0 and force it to simple 0.0.

On my machine, atan2 returns  -PI for the -0.0 case.

There are reports that microstation property browser show different angles for some matrix-to-rotations conversions.
If this atan2 behavior varies from machine to machine, that would explain it.