Fix scoping issue with static_assert in template. Minor doc fixes.
The scoping issues with the static_assert came from the use of std::is_base_of template method. The template is
resolved, sometimes, after the preprocessor has already tried to parse the static_assert. If this happens the
static_assert does not know the template types resulting in the '<>'s being ignored and the ',' inside of them
being extra arguments to the static_assert.

The fix for this is to add paranetheses around the template std::is_base_of call so that the static_assert can
assume that it is one parameter instead of multiple.