BF: Fix cache performance issue when caching ~100k+ instances for full persistance
Fix basically delays using of ECInstanceFinder only when it is needed.
For "*" or "$id" selects it would not be needed, so would not even occur.
BCCS usecase was always caching all data for each instance, so this fix moves
use of ECIntsanceFinder to location where it is not reached in such workflow.

Add example DataSourceCachePerformanceTests that are disabled by default.

TFS#811510
TFS#818264
SRB:Julius.Cepukenas