TFS#922840 - must allow all bridges to at least see all attached models. Cannot stop at first attachment that is not assigned to the bridge. That is because we have real world cases of master displine 1 attaches reference discipline2 attaches nested reference discipline1.
After initial model visiting functions run, we now call UnmapModelsNotAssigned to bridge, to remove
unassigned models from the BIM, syncinfo, and m_v8ModelMappings. The rest of the conversion should
see only assigned models ...

... but I also added belt-and-suspenders filters in all loops that look at m_v8ModelMappings and m_v8Files to
make sure the bridge does not process files that are not assigned to it. This is just in case unassigned
models are somehow discovered and added into the list again.