Enable logging for Android
the old logging system did not produce any logs on android devices.
now it does.

I can not directly expose the logger
#define LOG (*NativeLogging::LoggingManager::GetLogger(L"DgnCore"))
as someone would write LOG.debugv(...) with would cause errors on android.
Thus all logging is done through macros similar to DgnClientFx.

i have no idea what to do with isSeverityEnabled so it is defined as false on android.
On other platforms it works as it did althouth it is now accessed through macro.