VSTS#46995: Test failure: RulesDrivenECPresentationManagerContentTests.MergesPrimitiveArrayPropertyFieldsAndRowsOfDifferentClassesWhenSomeValuesAreNull
Can't simply swap two rapidjson::Documents - they only swap values, but not allocators. So basically item's Document was using allocator from another Document which got destroyed afterwards with all the data. The random factor came from the use of bmap in ContentProvider::LoadCompositePropertiesFieldValue which basically randomizes the order in which the nested content is loaded based on ECClass pointers.