Merged PR 13786: iModelGcs check fixed
My check that the GCS exists in the iModel was wrong. I should have checked for a valid pointer as well as for the validity of the GCS itself. There is an IsValid method on the smart pointer and an IsValid method on BaseGCS. The validity of the smart pointer needs to be checked before the validity of the GCS. This caused a crash where there was an iModel that had an ecefLocation but no GCS. That happens only for iModels created from reality data, so does not affect our EAP users.