Merged PR 14928: Must avoid incomplete navigation properties after element insert and update
When iModel.js does an insert and supplies JSON properties, we allow the caller to specify NavigationProperty values that contain only the target element ID and omit the relClassName. In that case, we would have to do an ECSql query in order to get the relClassName from the ec tables. To make sure that happens and to avoid having an incomplete NavigationProperty value in the cache, we discard the entire set of auto-handled property values from memory, so that the next request will be fulfilled by an ECSql query.

This problem can also crop up in an update, so we discard cached properties after updating, too.

Related work items: #66240, #66256