Merged PR 20589: yet another attempt at clean termination of folly ThreadPool
This will make your head hurt, but since this is exceedingly hard to debug more eyes will help.

There is a ThreadPool object that holds a condition variable that is used to synchronize with Worker objects that have a Work method that runs on another thread. Worker objects are reference counted and held in a vector on the ThreadPool object. When the ThreadPool is deleted, it tells the Worker objects to stop. Some ThreadPools are static objects, whose destructor is called by the system when exit() is called.

On Windows (optimized only, apparently) the system terminates the worker threads before it deletes static objects. So the pool cannot do anything in its destructor that relies on the other threads responding.

On Linux, the static destructors are called before the threads are terminated. They must tell the other threads to stop or the system hangs. But, the Pool object may be deleted before the other threads run.

Grigas, for your non-static Pool, if you want to make sure your other threads are "done" before you delete the pool, you should call "WaitForIdle" before you delete the pool.