Merged PR 26462: Add option to disable concurrent BRep facetting
Add option to disable concurrent BRep facetting

My testing has found that running PK_TOPOL_facet_2 in its concurrent
mode is always significantly slower (no matter how many threads are
attempting to run in it). I've reported this to Parasolid support and
they are tracking the issue as IR 8415939.

This change defaults concurrent facetting to enabled so there is no
change in the current behavior. I recommend changing this option to
false for tile generation, and placing that under a feature gate.

I am forcing this option to always disabled for ExportGraphics.