Merged PR 29811: Handle nested xRef files becoming out of sync with the root master file
Handle nested xRefs out of sync with the root master file: when a new DWG file
is attached to an existing xRef file without updating the master file, the newly
added attachment becomes out of sycn with the master file, VSTS#84381. We need
to cache these files and detect file changes accordingly.