Merged PR 35252: Update DgnElement::ToJson to use the RowFormat::IModelJs for JsonECSqlSelectAdapter
Update DgnElement::ToJson to use the RowFormat::IModelJs for the JsonECSqlSelectAdapter to be consistent with the queryRows() in iModel.js.

Fixes the IGeometry issue some of the Domain implementations are currently facing.

This gets us closer to the desired JSON format output in all cases from DgnElement.  There is still work to do to get an Element that is already in-memory to properly be returned and ensure there is consistency between ECSql and ToJson.