Merged PR 38353: The iModelJsNative.ts is now part of the imodeljs-native package
This is a step toward moving iModelJsNative.ts from imodeljs to imodel02. iModelJsNative.ts declares the imodeljs-native API.

This change also adds the beginnings of unit tests for the imodeljs-native package itself. These tests are written in TypeScript and use the addon's declared TS API. They call the native methods exposed by the addon. These are *not* mocha tests. They are simple functions written in TypeScript. So, running the tests is just a matter of running node.exe and passing it the test index.js file. That makes it easy to launch the tests under the debugger and set breakpoints in the addon native code.