Merged PR 39288: BeSQLite: Destroy app data objects after they're removed from the map
We don't want to appdata object to be destroyed while it's being removed from the map - in certain cases, where appdata's destructor is used for observing db close events, we had a situation where we could still find this object (with 0 ref count) in the map. To avoid it, we keep a ref while removing and the object gets destroyed when we return from this method.