Merged PR 39372: iModelBridgeFwk s/ look for assets in parent dir
The logic in iModelBridgeFwk that computes the default directory for the *framework's* assets directory was written back when a bridge dll and the iModelBridgeFwk.exe were in the lib/x64 subdirectory. Now, they are in the bin subdirectory. So, the logic to locate the framework's assets directory should now look up one directory, not two. To maintain backward compatibility. I kept the old logic. Now, the framework looks relative to the exe's directory, then relative to the parent directory, and then relative to the parent's parent.