Merged PR 44099: Remove loophole where iModel.js app could have modified an iModel without requesting locks and codes
If an iModel.js app opened an iModelDb in ReadWrite mode and never used the iModelDb.concurrencyControl object, it would have been able to make changes and push them without ever requesting locks or codes. A recent change to iModel.js backend should have put a stop to that. This PR is to add checking code to the addon so that such a loop hole would be detected if it were ever opened again in the future.