Merged PR 44268: V8Converter crash-recovery
Don't update the last-access-time property of the XSA that points to a RepositoryLink until all content from the v8 file has been processed. If conversion crashes part way through, the converter must be able to see that the file still needs to be processed. It will then use normal update logic to detect which models and elements were already converted and which did not get converted. It will convert the latter.

Abandon bulk update mode when unwinding and releasing locks after a crash. We must make sure that the lock release requests are sent to iModelHub and not get tangled up in a pending bulk request.

Related work items: #171870