Merged PR 44653: Fix missing face attached materials for instanceable geometry
If we decide not to try instancing a geometry part reference, e.g. because instancing is disabled or the instance is clipped by tile boundaries, we create an InstancedGeometry object. That thing forwards to GetPolyfaces(), which adjusts DisplayParams to account for face-attached materials. Later, when we try to resolve the DisplayParams to match the instance's symbology, we lose those face attached materials.
BF #152756

Related work items: #152756, #173876