Merged PR 44685: Failure to detect job subject lock denied error
This affects only the "new" briefcase manager.

When a desktop bridge job holds the exclusive lock on the job subject (channel parent), then a second job that tries to write to the same channel should fail right away with a clear locking error. It was not. Instead the fwk reported that the lock was obtained, and then the second job would run into all kinds of contention later.

I was confused when I first wrote the code. I did not seem to understand that RepositoryStatus::LockAlreadyHeld means that *another* briefcase holds the lock, not my briefcase!

Related work items: #173929