Merged PR 44841: More fixes for EWR and HS2
Some schema import fixes for EWR and HS2 - need to ignore another case in ECDb of a schema upgrade failure; need to check SchemaKeys in reverse minor version order when trying to locate a schema during de-serialization.

Need to delete duplicate rows in bis_ElementRefersToElements even though it should be impossible to have duplicate relationships in V8.  For HS2.