Merged PR 46458: EN: Look at the custom xml string to create the point cloud attachment.
We ran into a case where a point cloud attachment moniker did not have a provider Id information. In the debugger I found that the custom xml string contains the document moniker we need to resolve the filename So with this change we look at the custom xml string to create the point cloud attachment.

Related work items: #175178