Merged PR 46669: abandon bulk update mode when unwinding an error
Follow up to a recent change to abandon bulk update mode when unwinding an error. I forgot to trigger this feature in the case of a true crash. I have since learned that we must abandon bulk update mode when cleaning up in the case where the bridge or framework does not crash but does return early with an error.

Related work items: #177276