Merged PR 47322: Add support for IsNotSpatiallyLocated and IsPlanProjection on GeometricModel3d when present and safely skip when using an older schema
Add support for IsNotSpatiallyLocated and IsPlanProjection on GeometricModel3d when present and safely skip when using an older schema.

The goal was to not force a schema upgrade. This was accomplished by waiting until the actual DgnDb context was known before building the ECSQL statements. This allowed a check for the existence of a property on the ECClass before adding it to the ECSQL statement.

Related work items: #181848, #181850