Merged PR 47710: Default gtest BreakOnFailure=1 to fail reliably
I believe the logic of defaulting BreakOnFailure=0 when running gtest dates from DgnClientSdk when it was rigged to run all tests during the build as a first pass, but only fail / stop the build at the very end if any failed. We now (a) do not do this post-processing in general, and (b) can run tests in a separate build pass (supported by BB). Therefore, make the default to stop on failures immediately, and if you really want to try and ignore in a special case, you can opt-in to ignoring (vs. vice-versa) by defining BEGTEST_IGNORE_FAILURE.