Merged PR 47892: Fix minimum range detection.
We were attempting to detect the minimal (1mm to a side) element range in order to avoid omitting individual point primitives from tiles - but that detection was unreliable because RangeIndex converts the ranges to single precision.
Added a boolean to FBox to indicate whether the original double-precision element range was the minimal range.
Removed the category Id to mitigate the increase in sizeof(FBox) and because nobody uses it.
Removed ScanCriteria.
Was tempted to remove ViewController and DgnViewport but exercised restraint for now.

Related work items: #185197