Merged PR 48118: Add ECSqlStatement::TryGetParameterIndex which is the same as GetParameterIndex except that it does not log errors if the parameter is not found.
Add ECSqlStatement::TryGetParameterIndex which is the same as GetParameterIndex except that it does not log errors if the parameter is not found.

The goal was to not log errors if a missing parameter name was an expected condition. For example, the DgnModel.cpp case where the actual parameters depend on the version of the BisCore schema.