Merged PR 49574: Made error message more explicit as it caused confusion
In an email thread, Jason Chickneas mentioned he was confused by the old error message and spent some time checking his mixins. We decided to improve the error message so it is more concrete.
The new wording is up for discussion, we can adjust before closing the PR.

Related work items: #202258