Merged PR 49889: sheet-sheet attachments are invalid
Bug#29156:
Commit c2caa47cbca85d3836e62699646c8693572210cd on 8/22/2107 changed this logic to always call CopyEffectiveModelType.
That is incorrect in the case where the attached model is a sheet, since a sheet cannot be an attachment in an iModel.
The copy must become a drawing in that case. I cannot fathom when the copy should ever be anything but a drawing!?

I also tweaked the progress meter in iModelBridgeFwk. The "push" task is counted in the preset total task count.

Related work items: #29156