Merged PR 52795: Option to make bridge fail when it encounters a stale file
I was able to add the logic to recognize and apply this option to the V8Converter.
I could not do that for non-V8 bridges. I have enhanced iModelBridgeSyncInfoFileChangeDetector to allow a bridge to detect a stale item. The bridge can then, optionally, return an error in that case..
(iModelBridgeSyncInfoFileChangeDetector does implement the option to ignore stale items, so bridges don't have to do anything to get that optional behavior.)
I also fixed a bug in the command-line processing logic for the new --fwk-ignore/error-on-stale-files options.

No impact on existing bridges and BAS jobs. Stale file rejection is optional, and the default is not to apply any special logic to back-dated files, as we have always done.

Related work items: #203143