Merged PR 54609: VSTS#215468 - Need to check stored schema and not incoming V8 schema for the
Not all V8 domain schemas have the DynamicSchema custom attribute set on them.  The bridge automatically adds this CA to all imported schemas.  Therefore, when looking for the CA during an update, we need to check the stored schema and not the incoming schema to see if it is dynamic.