Merged PR 57175: ok to write to submodel of RepositoryLink any time
Many bridges and subsystems write to submodels of RepositoryLinks elements that live in the special RealityDataSourcesModel during the data conversion phase (normal channel). This is shared data in a shared model, and so such updates really should be done in the shared channel. However, it is too late to enforce that constraint now. There is too much code that is breaking the rule. That code is owned by many different groups. There is no straightforward way to refactor it.
The consequences of allowing concurrent updates to submodels of elements in the RealityDataSourcesModel should be relatively benign. In the very worst case, if two different bridges see the same reference to the same reality data, both will try to create the same RepositoryLink element. The second will fail with a code-reservation error and will have to be re-run.

Related work items: #222334