Merged PR 57278: Improve tile elision
Improve detection of empty sub-volumes during tile generation so that we can elide requests for child tile content for child volumes known to intersect no geometry.
Previously we always tested for intersection of the geometry's AABB with the tile's AABB. This is a very coarse test that can result in pathological cases in which we infinitely sub-divide empty volumes.
The changes allow for tighter intersection tests for particular types of geometry.
The following common cases are handled:
- Line strings (e.g. road/railways): intersect each segment.
- Polyfaces (e.g. concept station bridges/roadways): intersect each triangle.
- Box solid primitives (e.g. huge green cubes that some of our customers use to denote volumes of the model): intersect faces.
- Planar regions: intersect bounding rectangle.
- Point strings: intersect each point.

We initialize a set of "empty" flags for each sub-volume of the tile volume (4 for 2d, 8 for 3d) as well as a flag indicating that all sub-volumes have been identified as non-empty.
For each geometric primitive that intersects the tile volume:
  If we've already established that all sub-volumes are non-empty; or the geometry's AABB is not large enough to bother with (at least 20% of the tile's volume); or the geometry is not one of the types listed above:
  - Just do the AABB-AABB test against each empty sub-volume.
  Otherwise:
  - Identify empty sub-volumes that intersect the geometry's AABB.
  - Run the intersection test against each.
  - As soon as an intersection with a given sub-volume is identified, remove it from the set of candidates.
  - As soon as the set of candidates is empty, stop.

I'm sure there are plenty of inefficiencies. e.g., there's certainly a cheaper way to test quad-range intersection (I'm currently doing it as two triangle-range intersection tests).

Related work items: #102543, #220011