Merged PR 57720: ECDb was not calling ResetInstanceIdSequence on briefcaseId assigned
As a result, a restored briefcase would have the linktable relationship id sequence reset to briefcaseid + 0. If there were in fact relationship instances in the iModel already, all subsequent attempts to insert new ones would fail with a unique key constraint.

Also, there are a number of problems with the logic to report outlier elements. The query for the element synchinfo was not correct. Only one (the last) of many was being reported, despite the loop. That may have been intentional, since we may not have wanted to generated a string of unbounded length. In any case, I renamed the function and eliminated the loop to make it clear that only one was being reported. If we want to report them all, we should write individual issues for each, rather than generating a huge string that describes them all.

Related work items: #223469