Merged PR 59984: Allow JS to manage locks and codes
This is a step toward moving all responsibility for managing locks and codes into iModel.js. Currently, JsInteropBriefcaseManager.cpp implements a C++ BriefcaseManager to manage locks and codes. iModel.js on the master branch relies on this and works with it. This change is a transitional step. It allows iModel.js to effectively turn off the JsInteropBriefcaseManager. This capability will be used while developing and testing the iModel.js side. When the changes to iModel.js are pushed and it can take over management of locks and codes, a follow-up change will be required to remove BriefcaseManager from JsInteropBriefcaseManager.cpp and all of the related functions that work with native resource requests.

Note that this has nothing to do with the C++ BriefcaseManager that is used by native-code bridges. That is a separate implementation of IBriefcaseManager and will not be affected by the work on iModel.js. It will remain in place even after JsInteropBriefcaseManager is removed.