Merged PR 62472: Fix bugs in iModelBridgeSyncInfoFileChangeDetector
Fixes for geobridge:
Removed the special implementation of ChangeDetector that tried to optimize the case of the initial conversion of a given joblet. It caused a bug in geobridge (and in the ToyTile example bridge) in the case where iModelBridge::RecordDocument is called twice in the initial conversion, once in _InitializeJobSubject and again in _ConvertToBim. A bridge should make both calls (the ToyTile example shows this). The special change detector, however, was causing a duplicate RepositoryLink element to be created, leading to the situation where _DeleteElementsNotSeenInScopes would delete all of the elements that were scoped to the first RepositoryLink element. The solution is to use the real change-detector all the time. It correctly detects when the RepositoryLink element already exists and does nothing.
Added a second overload of RecordDocument to allow the caller to detect when the file's last-modified time is unchanged.
Fixed a huge bug in iModelBridgeSyncInfoFile::ChangeDetector, where we were assigning the wrong ID to the so-called "ROWID" property of the returned conversion results. That was causing a bug when a bridge was run in update mode. I also deprecated this property and added a new property to capture the ExternalSourceAspect's instance ID.
SAC adapter (test harness) was failing to call SaveChanges after calling bridge->_MakeSchemaChanges.