Merged PR 63032: Missing syncInfoRecord in Changed case
Another case where tiModelBridgeSyncInfoFile::ChangeDetector::_UpdateBimAndSyncInfo was returning incorrect results. This surfaced in geobridge.
Added some comments to help other developers understand what the various structs and member variables are supposed to represent. The names are confusing, because they are based on the old design, where we had a separate "syncinfo" file. Hopefully, the comments will help to explain how they correspond to the current design based on ExternalSourceAspects.