Merged PR 67577: signal-handler; channel rules exception
Add signal-handler to iModelBridgeFwk to make sure that locks are released when process is killed. This is done by iTwin Synchronizer.
Add special case for PresentationRules in channel rules. This is a top-level model that bridges create when rules are needed, e.g., when IFC schema is imported.
Add MstnBridgeTest to verify that .ifc can be processed (with and without channel rules). Import of IFC schema is what causes PresentationRules model to be created.