Merged PR 69416: IModelJSON reader for cone use explicit orientation from input json object (rather than infering for possibly ambiguous points)
(For Shaun)
The imodeljson reader for a cone did not look for orientation data in the source object.
This works fine for the usual case of a cone that has circular sections in the plane perpendicular to the line
   between top and bottom centers.
A bim transformation app encountered "cones" that have identical start and end centers, i.e. a annulus between two circles in
  the plane.   This makes it incorrect to look for cone axis as line between the (identical) points.
The reader now looks in the json source for orientation to resolve that.