Merged PR 68808: sqlite_3.31.1
Update sqlite to latest. This time update from trunk as there are not many changes since 3.30.

Released notes: https://www.sqlite.org/changes.html

**Page cache size**
The cache size is now back to 32 MB from 64 Mb. It seem it cost us potentially 576 (9 connections *64 Mb each) Mb at worse on backend. Though it hardly ever happens. The cache extended for shared-cache which need a bit testing before we enabling it. It drastically reduce memory required by each backend to a single cache-size. But as of now we are unable to test it.