Merged PR 70606: Delete References Subject when V8 attachment goes away
A V8 reference attachment is represented in the iModel by a Subject element with JSON properties that identify it as a "Reference" to the model. The only purpose of this Subject element is enable the presentation tree to show a view of the iModel's contents that look like the V8 reference attachment hierarchy. When an attachment is removed in V8, the user expects this Subject element to be deleted from the iModel and thus disappear from the presentation tree. That was not happening.

To clean up abandoned references, first of all, we have to start tracking them - they need ExternalSourceAspects. Element deletion is driven by comparing XSAs with elements seen.

Deleting a References Subject is complicated by a flaw in the underlying design, where the Subject is not just a reference to a model but actually its parent. That is, at least one of the many References Subject elements is. We are working around that for now.

The re-parenting logic may also move a model to a different Job Subject hierarchy. As a result, another bridge may end up "owning" the common reference. That violates our channel rules. Must live with this for now.

Related work items: #278982