Merged PR 71998: Search embedded refs for sheets and drawings
Normally it's up to the user to identify files that contain sheets and drawings outside of the files that define 3D models. In the case of a i.dgn file, however, the user has no way to specified embedded reference files. The V8 publisher will include such reference files in the package, but the V8 converter will not follow reference attachments from 3D into 2D. So, this change compensates by automatically adding all embedded files to the list of files to search for sheets and drawings.

Related work items: #284079