Merged PR 71735: delete orphan reference models
This cleans up PhysicalPartitions that are left after their parent References Subjects are deleted. This special logic is for the case where the normal DetectDeletedEmbeddedFiles cannot detect the abandoned reference files because a iModelBridge_MatchOnEmbeddedFileBasename recipe has been set such that new and old reference files have the same *unique* names, so that the normal detector thinks that the old refs are still in the package files. As a work-around, this new function detects PhysicalPartitions that were abandoned by their parents and deletes them.
This change also adds logic to make sure that PhysicalPartitions are not mistaken for orphans when they are in fact referenced. An abandoned PhysicalPartition is now picked up and reparented by the first new reference that may come along. This happens if a ref is dropped in one file and then added in a second file (in order of processing).