Merged PR 72566: allow bridge to change class of existing element
The IFC bridge  needs to change the ECClass of an existing, previously converted element. This must be done as a delete + insert. The bridge itself must be changed to detect this case and do the delete. The iModelBridgeSyncInfoFIle class had to be changed to recognize when an update is really an insert.
Also fixed failing project extents tests.

Related work items: #285859