Merged PR 74318: Fix while preventing dynamic classes from being deleted
Preventing dynamic classes from being deleted after an update. This can happen in a CIF dataset if a featureDef is not used by any element anymore after an update. The associated dynamic class for that featureDef will not be regeneated and therefore, could be deleted. However, that's considered as a major schema update and it is not supported. Keeping those dynamic classes alive to prevent that issue.

This fix is also needed so I can point platform to a test-case reproducing the situation of an element changing its classId. Such case needs to be supported by them - deleting it and recreating it under its new classId.