Merged PR 75542: name referenced models according to V8 naming rules
Spatial model names are now showing up in the UI by themselves, i.e., not in the context of the reference attachment hierarchy. The clash detection is an example. Users at WSP requested that the name of a spatial model created by a bridge should identify the source file that contains the model.
Based on that user  request, Glen Worrall and others suggested that the bridge should name a spatial model after the reference attachment that was used to find it, that is, ref logical, filename, modelname. If the model is referenced multiple times, the name should be the model name, qualified by the filename.

Related work items: #291480