Merged PR 76033: Remove use of ECUtils from the Native Addon
I stripped ECUtils and replaced the code that was still using it.
Also made some methods in ECJsonUtilities take Utf8CP instead of Utf8String* and take const pointers to properties to avoid some gross const_casts.
There is a second PR for imodeljs coming after this one, which adds more roundtrip tests for the code that was changed here.

Related work items: #269229