Merged PR 76393: avoid stale pointers when forwarding args to bridge
iModelBridgeFwk parses the command line in multiple steps. At each step, some arguments are recognized and others are forwarded to the next step. This change ensures that the arguments to be forwarded are stored long enough to be forwarded. Previously, fwk was relying on the raw string pointers persisting long enough for this to happen. The recent switch to std::vector and std::string invalidated that old design. The new design removes all doubt by saving and forwarding copies,

Related work items: #289787