Merged PR 76733: Update RepositoryLink userLabel and URL
When ProjectWise doc URN/GUIDs or an embedded file recipe is used to identify files, the name of a file can change without changing the file's identity. As long as the file's identity is the same, the converter will map it to the same RepositoryLink element. The RepositoryLink element's UserLabel property is based on the filename, so it must be updated when the file is renamed. It is theoretically possible for the file's URL to change as well, although that cannot happen in ProjectWise case.

This change affects all bridges that call iModelBridge::UpdateRepositoryLinkDocumentProperties.

Related work items: #294621