Merged PR 77101: VSTS #296785 - When merging schemas, disregard version number when checking to
VSTS #296785 - When merging schemas, disregard version number when checking to see if schema is already referenced.

The converter is merging the existing schema with the incoming schema.  These two versions can potentially have the same referenced schemas, but with different versions.  The current logic allows adding both versions to the list of referenced schemas, which is not acceptable to ECDb.  Therefore, when looking to see if the schema is referenced already, disregard version.

Related work items: #296785